Fixed an issue with invalid heights.

+++ main.py
@@ -1,6 +1,7 @@
 import sys
 from pathlib import Path
 import xmltodict
+import time
 from PIL import Image
 from math import isclose, ceil
 from sheet_entry import SheetEntry
@@ -21,6 +22,7 @@ EXPORTS_WARNING_PATH = EXPORTS_PATH.joinpath('__DO NOT SAVE HERE__.txt')
 # {'property': 'numFrameEntries'}
 
 def main():
+    time_begin = time.time()
 
     if EXPORTS_PATH.exists():
         delete_directory(EXPORTS_PATH)
@@ -32,13 +34,14 @@ def main():
             print('Do not save any changes to this directory, it will be overwritten by the next operation!',
                   file=warning_file)
 
-    existing_sheets = []
+    existing_sheets = {}
 
     if RESOURCES_EXISTING_SHEETS_PATH.exists():
         for sheet in RESOURCES_EXISTING_SHEETS_PATH.iterdir():
-            if sheet.is_dir():
-                continue
-            existing_sheets.append(sheet.stem)
+            if sheet.is_file() and str(sheet).endswith('.png'):
+                sheet_image = Image.open(sheet)
+                existing_sheets[sheet.stem] = int(sheet_image.height / 256)
+                sheet_image.close()
 
     sheets = {}
 
@@ -62,19 +65,36 @@ def main():
 
     print('\n', end='')
 
+    next_bar = 0
     for key, value in sheets.items():
+        height_minimum = 0
         if existing_sheets:
-            if key not in existing_sheets:
+            if key in existing_sheets.keys():
+                height_minimum = existing_sheets[key]
+            else:
                 new_sheets.append(key)
-        clone_sheet(key, value)
-        print('.', end='')
+        clone_sheet(key, value, height_minimum)
 
-    print('\nDone! {} sheets stitched'.format(sheet_count))
+        if next_bar == 4:
+            print('|', end='')
+            next_bar = 0
+        else:
+            print('.', end='')
+            next_bar = next_bar + 1
 
     if new_sheets:
         print('New Sheets:')
         for sheet in new_sheets:
             print('\t{}'.format(sheet))
+
+    time_total = time.time() - time_begin
+
+    if time_total < 60:
+        time_total = '{} seconds'.format(time_total)
+    else:
+        time_total = '{}:{} minutes'.format(int(time_total / 60), int(time_total % 60))
+
+    print('\nDone! {} sheets stitched in {}'.format(sheet_count, time_total))
 
 def consume_sheet(pack_name, root, sheets):
     root = root['object']['void']
@@ -114,7 +134,7 @@ def delete_directory(target_path):
             sub.unlink()
     target_path.rmdir()
 
-def clone_sheet(sheet_name, entries):
+def clone_sheet(sheet_name, entries, height_minimum, cleanup=True):
     sheet_path = EXPORTS_PATH.joinpath('{}.png'.format(sheet_name))
     sheet_assets_path = EXPORTS_PATH.joinpath(sheet_name)
     sheet_assets_path.mkdir()
@@ -139,19 +159,25 @@ def clone_sheet(sheet_name, entries):
         else:
             cell_height = ceil(index_last / 8.0)
 
-    cell_height = int(cell_height)
+    cell_height = max(int(cell_height), height_minimum)
 
     stitch_asset(sheet_assets_path, sheet_path, index_last, cell_height)
 
-    delete_directory(sheet_assets_path)
+    if cleanup:
+        delete_directory(sheet_assets_path)
 
 def clone_asset(export_path, entry):
+    if entry is None and export_path.exists():
+        return
+
     result = Image.new('RGBA', (128, 256))
     if entry is not None:
         source = Image.open(entry.asset_path)
         result.paste(source, (entry.offset_x, entry.offset_y))
+        source.close()
 
-    result.save(export_path)
+    result.save(export_path, 'PNG')
+    result.close()
 
 def stitch_asset(source_path, export_path, count, cell_height):
 
@@ -164,6 +190,7 @@ def stitch_asset(source_path, export_path, count, cell_height):
         for x in range(0, 8):
             source = Image.open(source_path.joinpath('{}.png'.format(index)))
             sheet.paste(source, (x * 128, y * 256))
+            source.close()
             index = index + 1
             is_done = count <= index
             if is_done:
@@ -171,7 +198,8 @@ def stitch_asset(source_path, export_path, count, cell_height):
         if is_done:
             break
 
-    sheet.save(export_path)
+    sheet.save(export_path, 'PNG')
+    sheet.close()
     # print('source: {}, export: {}, count: {}, cell_height: {}'.format(source_path, export_path, count, cell_height))
 
 if __name__ == '__main__':
